# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: CheeSwhiz
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/plain
paths:
  /cheese:
    # binds a127 app logic to a route
    x-swagger-router-controller: cheeSwhiz
    get:
      summary: Get all the cheeses that CheeSwhiz has to offer.
      description: Returns all cheeses in the database.
      # used as the method name of the controller
      operationId: allCheese

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CheeseArray"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      summary: Add a cheese to our database!
      description: Have a cheese that we don't? Share your expertise with us!
      # used as the method name of the controller
      operationId: postCheese
      parameters:
        - name: Cheese
          in: body
          description: The cheese to create.
          required: true
          schema:
            $ref: "#/definitions/CreateCheese"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/Cheese"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /cheese/{id}:
    x-swagger-router-controller: cheeSwhiz
    patch:
      summary: Update a cheese!
      description: A super user/admin may make updates to the cheeses
      # used as the method name of the controller
      operationId: updatedCheese
      parameters:
        - name: id
          in: path
          description: cheese id to update
          required: true
          type: integer

        - name: Cheese info and Token
          in: body
          description: The cheese to update.
          required: false
          schema:
            $ref: "#/definitions/UpdateCheese"

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UpdatedCheese"
        # responses may fall through to errors
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
            
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"

        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      summary: Delete one of our cheeses!
      description: A super user/admin may delete a cheese from our database.
      # used as the method name of the controller
      operationId: deleteCheese
      parameters:
        - name: id
          in: path
          description: cheese id to delete.
          required: true
          type: integer

        - name: Super User Token
          in: body
          description: The super user's authentication token to authorize the deletion of a cheese.
          required: true
          schema:
            $ref: "#/definitions/SuperUserToken"

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DeleteCheese"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /cheese/random/{number}/{animal_or_firmness}:
    # binds a127 app logic to a route
    x-swagger-router-controller: cheeSwhiz
    get:
      summary: Find a random amount of cheeses.
      description: Find a random number of cheeses by some parameter from our database.
      # used as the method name of the controller
      operationId: randomCheeseGenerator
      parameters:
        - name: number
          in: path
          description: Enter the number of random cheeses you would like.
          required: true
          type: integer

        - name: animal_or_firmness
          in: path
          description: Enter a either animal or firmness based on what parameter you would like to return cheeses by.
          required: true
          type: string

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CheeseArray"
        # responses may fall through to errors
        "400":
          description: Bad Request
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: Not Found
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /user/login:
    # binds a127 app logic to a route
    x-swagger-router-controller: login
    post:
      summary: Log a user into CheeSwhiz.
      description: Log a user into CheeSwhiz so that they can access their favorites.
      # used as the method name of the controller
      operationId: logInUser
      parameters:
        - name: credentials
          in: body
          description: User's email and password they are logging in with.
          required: true
          schema:
            $ref: "#/definitions/userRegLogin"

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/loggedIn"
        # responses may fall through to errors
        "400":
          description: Email and Password didn't match.
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  CheeseArray:
    description: An array of cheeses based on the user's selection.
    required:
      - cheeses
    properties:
      cheeses:
        type: array
        description: An array of cheeses
        items:
          $ref: "#/definitions/Cheese"
  Cheese:
    description: A single cheese in the database.
    required:
      - name
      - firmness
      - animal
      - id
      - user_id
    properties:
      name:
        type: string
        description: The name of the cheese.
      firmness:
        type: string
        description: The firmness of the cheese.
      animal:
        type: string
        description: The animal milk it is made of.
      id:
        type: integer
        description: Primary key of the cheese.
      user_id:
        type: integer
        description: The user's id who added the cheese to the database.
  CreateCheese:
    description: Adding new cheese to the databse.
    required:
      - name
      - firmness_id
      - animal_id
    properties:
      name:
        type: string
        description: The name of the cheese.
      firmness_id:
        type: integer
        description: The firmness id of the cheese.
      animal_id:
        type: integer
        description: The animal milk id it is made of.
      user_id:
        type: integer
        description: The optional user_id. Can be null.
  UpdateCheese:
    description: Updating a cheese in the database. Please update name, firmness, or animal.
    required:
      - token
    properties:
      name:
        type: string
        description: The name of the cheese.
      firmness_id:
        type: integer
        description: The firmness id of the cheese.
      animal_id:
        type: integer
        description: The animal milk id of the milk it is made of.
      token:
        type: string
        description: A users authentication token
  UpdatedCheese:
    description: The cheese information that has been updated.
    properties:
      name:
        type: string
        description: The name of the cheese.
      firmness:
        type: string
        description: The firmness of the cheese.
      animal:
        type: string
        description: The animal milk it is made of.
      id:
        type: integer
        description: The id of the cheese to update.
  DeleteCheese:
    description: The cheese that is being deleted.
    required:
      - name
      - firmness
      - animal
      - id
      - user_id
    properties:
      name:
        type: string
        description: The name of the cheese.
      firmness:
        type: string
        description: The firmness of the cheese.
      animal:
        type: string
        description: The animal milk it is made of.
      id:
        type: integer
        description: The id of the cheese to update.
      user_id:
        type: integer
        description: The id of the user who created the cheese.
...
  userRegLogin:
    description: Used for either creating a user or logging a user into CheeSwhiz.
    required:
      - email
      - password
    properties:
      email:
        type: string
        description: The user's email address in the database.
      password:
        type: string
        description: The user's password or desired password.
...
  SuperUserToken:
    description: A super user's (or admins) authentication token for CheeSwhiz.
    required:
      - token
    properties:
      token:
        type: string
        description: The super user's access token.
...
  loggedIn:
    required:
      - email
      - id
      - token
      - super
    properties:
      email:
        type: string
        description: The users email.
      id:
        type: integer
        description: The users id.
      token:
        type: string
        description: The users validation token sent in the body.
      super:
        type: boolean
        description: If the super user is a super user
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string